+++ test_app/config/initializers/ultra_settings.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+UltraSettings.runtime_settings = SuperSettings
+UltraSettings.super_settings_api_path = "/super_settings"